Add main menu bar
Currently Save, Quit, and Quit w/o saving work properly. For Quit a Quit method
was added to OS.

Fixed a few potential crashes such as mesh changing causing the vertex buffer to
not be recreated at the correct point, and one where the focused window was being
used but not checked.