Added the ability to set an offset for position and rotation
These can be used for center of mass or graphical offsets.

Removed scale from view calculations as it isn't meaningful there.

Physics now uses a copy constructor when updating position and orientation to
ensure offsets are copied forward without having to pollute physics with knowledge
of the offsets.

Updated assets.