Refactor of types.hpp
We use proto::Component enum values as TypeID value. This make more easy to
follow the track to all, and make more simple to register new component types.
The price is that .protfile component name must be the same that the struct/class
name.

Also, we now have a runtime map of type ID to const char* . This is necesarry
to map stuff to Lua.