Fixed in/out of CollisionBody components to use it's new_shape property
Physics system might not get a chance to update it's internal shape and therefore
the components current shape after the component's new shape is set. We want to save
what it will be and not what it was.

camera_controller in main.cpp is now a pointer type to allow checking if it is valid.

Added the ability to show/hide GUI windows.