First step for reducing use of shared_ptr
Also simplified CollisionBody to now be simply a collision shape and motion_state.
The mass and both disable_X members are used when creating the rigid body and will
hopefully be refactored into something else.

This is being done in steps to ensure nothing breaks with a massive sweeping set
of changes of removing shared_ptr in place of raw pointers.