Fixed OBJ loader not working with object groups
Also when parsing faces each time usemtl was encountered a new face group was added.
This wasn't necessary and created a lot of lag when rendering. Now when usemtl is
encountered, a search is done on the previous material groups and the following faces
are added to that group if one was found.

Updated assets

Changed default vertex color from green to black.