Added constexpr to GetTypeX functions
Fixed a copy and paste typo in physics and changed scoping of motion state property updates.
Also added a check for a null collidable possibility.

Fixed appveyor to use upstream vcpkg.

Entity::Get now returns const

A change was required to selene and as such is only in my fork therefore changed submodule

Removed last of the travis-ci files.