Fix tautological comparison
The comparison before this change was doing nothing, comparing
this->entity_id with itself. It was misleading because it looked like it
was comparing entity_id with this->entity_id, but they're the same
thing. Looking at the surrounding code, this->entity_id should be
compared with data->entity_id.