Rename server's SERVER_PORT to PORT
Now that the tests compile in both the client and the server, using the
same name (tec::networking::SERVER_PORT) for two different variables
with different types is an ODR violation. Before this, it was okay,
because the client and server were separately compiled and never
compiled together.

An alternative to renaming shared names as they are discovered is to put
the client in tec::client and the server in tec::server, with common
remaining in the tec namespace.