General cleanup for warnings and style (#142)
* General cleanup of code, renaming variables, etc.

* Modify stb_vorbis.c to get rid of warning

Getting in a habit of doing this probably isn't a good idea, but this is
a purely cosmetic change just to fix a warning, so it shouldn't be a
problem.

* Add/improve logging to/in do-nothing catch blocks

Catch blocks that do nothing should at least print some kind of log
message. Should these re-throw, or is doing nothing correct?

* Fix tautological comparison

The comparison before this change was doing nothing, comparing
this->entity_id with itself. It was misleading because it looked like it
was comparing entity_id with this->entity_id, but they're the same
thing. Looking at the surrounding code, this->entity_id should be
compared with data->entity_id.

* Rename server's SERVER_PORT to PORT

Now that the tests compile in both the client and the server, using the
same name (tec::networking::SERVER_PORT) for two different variables
with different types is an ODR violation. Before this, it was okay,
because the client and server were separately compiled and never
compiled together.

An alternative to renaming shared names as they are discovered is to put
the client in tec::client and the server in tec::server, with common
remaining in the tec namespace.