RenderSystem refactor part 3
Adds shader type infrence, which looks up all the uniforms and
attributes the shader uses, engine sampler channels are typically set to
a constant, so this new system looks up those binding constants from a
list and sets them the first time a shader is used.
Adds in-engine shader include pre-processing, with debug line correctness
this includes changes to the Render config file to specify what can be
included.
* Adds RenderSystem commands to the console, additionally all console
  commands now take a std::string instead of whatever the hell they
  were doing before.
* Added "options" to the RenderSystem, enabling runtime switching of core
  shaders, and toggling debug view on/off via command.
* Add Emission to the GBuffer
* Converted the FinalPass to a shader, this also handles emissive.
* Add a parameter texture to the floor
* Changed Material "activation", now the render system binds material
  textures directly, providing a default in case a shader needs it.
* View and Model render transforms moved to quaternion + position.
* The lighting bounds function was making the radius too big, fixed!