remove non-entity On method
"On" came in two flavors, one with an entity ID and one without.
Everything used "On" without an entity ID, including events which... act 
on an entity ID, and then include said ID as part of the event anyways.
Switching to a single definition removes the need for a bunch of 
redundant "using" lines.
This also means a now single code path for events... which actually gets 
used.
Each declaration must include the eid term, even if you don't need it.
Hopefully it makes sense when and where an entity ID needs to be Emitted 
and used.