Refactor of filesystem code
FilePath renamed to Path, because it's more than just files.
Paths are now always in a generic form, so toGenericString was removed
GetNativePath will now decode the path and generate the native form.
Paths now include a "device", which is kind of like a URL scheme,
If the device is a drive letter, then the Windows version will use it,
Specifying an unknown device throws an exception.
The only engine defined device is "assets:" which represents the root
of the assets directory, all other device names are considered invalid
(except drive letters on Windows)
Resource and component loaders were changed over to the new system.
Also Path now includes commands to open files and return either a
C-style FILE or a C++ style fstream.
These changes will be essential to supporting virtual file access