Path constants and formatters
Replaced the GetAssetPath method with a constant.
Also added constants for scripts and shaders.
Added a formatter so paths can be specified directly to spdlog.
NormalizePath() and GetNativePath() were made private, this helped
find a few places where I needed to use OpenStream()
fixed OpenStream() to actually be able to write files.
PATH_OPEN_FLAGS are a bit flags field, but it wouldn't compile,
so I added some operators to deal with that...
Added Path::Relative as a shortcut for Subpath(1)
fixed a warning with vorbis-stream and a signed buffer size.